migrate NotebookApp into components
