+++ applications/play/components/Main.js
@@ -7,7 +7,7 @@ import CodeMirrorEditor from "@nteract/editor";
 import { BinderConsole } from "./consoles";
 import { Display } from "@nteract/display-area";
 import { KernelUI } from "./kernelUI";
-import { Outputs } from "@nteract/core/components";
+import { Outputs } from "@nteract/core";
 import { connect } from "react-redux";
 import { actions } from "../redux";
 import objectPath from "object-path";
