+++ packages/core/__tests__/providers/notebook-spec.js
@@ -7,7 +7,7 @@ import { shallow, mount } from "enzyme";
 import renderer from "react-test-renderer";
 
 import { displayOrder, transforms } from "@nteract/transforms";
-import { NotebookApp } from "../../src/providers/notebook-app";
+import { NotebookApp } from "../../src/components/notebook-app";
 
 import { dummyStore, dummyCommutable } from "../../src/dummy";
