+++ packages/core/src/providers/index.js
@@ -1,5 +1,4 @@
 // @flow
-import NotebookApp from "./notebook-app";
 import Styles from "./styles";
 
-export { NotebookApp, Styles };
+export { Styles };
